Moved Utility into own folder
