+++ DreamScape/DreamScape/Data/Utility/Helper.cs
@@ -8,7 +8,7 @@ using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
-namespace DreamScape.Data
+namespace DreamScape.Data.Utility
 {
     class Helper
     {
